requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
fastapi==0.114.2
greenlet==3.1.0
h11==0.14.0
idna==3.10
jwt==1.3.1
passlib==1.7.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
